+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Versioning follows a `MAJOR`.`MINOR`.`PATCH` pattern, where:
 ### Changed
 
 - Most configuration options have moved from env variables to the app config object
+- Hide Metadata resource type unless user has organization prefix
 
 ## [2.0.0](https://github.com/spraakbanken/mink-frontend/releases/tag/v2.0.0) (2026-06-01)